Implement initial passthrough to CITI's API
Endpoints are defined in `config/resources.php`. We assume that the
endpoints we offer in this service's API are the same as those offered
by CITI's API. See `routes/api.php` for how the config file gets used
to generate routes.

`PassthroughController` uses the endpoint-to-transformer bindings in
`config/resources.php` to run the right transformer for each endpoint.
Currently, there are no endpoint-specific transformers, but that will
soon change. The main function of `AbstractTransformer` via concrete
`GenericTransformer` is to transform each object returned from CITI's
API into an instance of `App\Transformers\Datum`. This class is a good
place for reusable data cleanup macros, but right now, its main purpose
is to loop over each field of each object, check if it's a string, and
null it out if it's empty, after trimming.

We composer require'd Guzzle for querying the upstream API. It's used
in `PassthroughController` via dependency injection, as defined in the
`GuzzleServiceProvider`.