Fix zero id fields in Artwork subdatums
All relationship fields should be null in cases where they point at
nothing. Here, they were defaulting to zero.

Additionally, there was an issue wherein arrays of objects weren't
being converted to Datum instances. This should be fixed with the
new magic methods.

Lastly, I'm not sure if we should standardize on empty arrays
returning empty, or null. Currently, we return empty in most
cases, but it's not completely consistent. Sync expects empty,
so I'm standardizing around that for now, but we should revisit.